refactor: init logic replaced into hook useInit
