+++ src/App.tsx
@@ -1,29 +1,11 @@
 import './app.module.scss';
 import { NavBar } from 'components/NavBar/NavBar';
 import AppRouter from 'router/AppRouter';
-import { useEffect } from 'react';
-import { onAuthStateChanged } from 'firebase/auth';
-import { auth } from './firebase';
-import { useAppDispatch, useAppSelector } from './hooks/redux';
-import { isAuth } from './redux-toolkit/reducers/isAuthSlice';
 import { Loader } from './components/ui/Loader/Loader';
+import { useInit } from './hooks/useInit';
 
 function App() {
-  const dispatch = useAppDispatch();
-  const { initialize, login } = isAuth.actions;
-  const { initializeSuccess } = useAppSelector((state) => state.auth);
-  useEffect(() => {
-    onAuthStateChanged(auth, (user) => {
-      if (user) {
-        const { uid } = user;
-        dispatch(login(true));
-        console.log('uid', uid);
-      } else {
-        console.log('user is logged out');
-      }
-      dispatch(initialize(true));
-    });
-  }, [dispatch, initialize, login]);
+  const initializeSuccess = useInit();
   if (!initializeSuccess) {
     return (
       <div className="mainPageLoader">
